requirements.txt
pydantic
langchain
langchain-openai
langgraph
uvicorn
ipykernel
python-dotenv
fastapi